Feature/#81 integration

+++ src/main/java/com/team5/WalkingWithWorld/walkingPaths/dto/WalkingPathsDTO.java
@@ -12,8 +12,8 @@ import java.util.List;
 @Setter
 @ToString
 public class WalkingPathsDTO {
-    private int id;
-    private int usersId;
+    private Long id;
+    private Long usersId;
     private String title;
     private String addr;
     private LocalDateTime createdAt;
